`Refactor CurrencyContext to CurrencyContext/currency, create separate provider for Antd and RootProvider`
